[if] deprecate `Cache`, lift `Session`s to `STuple`s instead
`Cache` is dangerous when type of a session changes.  The Codegen procedure
still reads to old session information instead of re-resolve to get the new
info.

For a `Cache` to be usable, it needs to have attributes like ReadOnly, WriteOnly
and ReadWrite.  Updates to the cache need to be systematic, which seems to be
hard to enforce at the type level.