[spec] change spec formats
Now `Nor` and `Had` partitions may only contain one range.
A list of specifications are allowed for each state specification to describe
the conjunction of facts.

For example, one may use

```haskell
{ q [0 .. 10] : nor
  ↦
  [ ⊗ i ∈ [0..5] . 0
  , ⊗ i ∈ [5..10] . 1
  ]
}
```
to specify fine-grained properties of *one* locus. The internal representation
of this locus is still a single sequence of nat for kets, and it should not be
interpereted as a locus of two ranges.